Deleted old reports implementation
